// b-nextjs-frontent-main/src/payload-types.ts
export interface ReviewCard {
  id: string;
  name: string;
  shortText: string;
  externalLink: string;
  images: string | AuctionMedia;
  countStars?: number | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "auction-media".
 */
export interface AuctionMedia {
  id: string;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "catalog-car".
 */
export interface CatalogCar {
  id: string;
  images?: (string | AuctionMedia)[] | null;
  auctionList?: (string | null) | AuctionMedia;
  externalId: string;
  lot: number;
  auction: string;
  brand: string;
  body?: string | null;
  model: string;
  modelDisplay: string;
  modelSlug: string;
  year: number;
  wheelPosition?: ('right' | 'left') | null;
  transmission?: string | null;
  enginePower?: number | null;
  horsepower?: number | null;
  driveType?: string | null;
  rating?: string | null;
  price: {
    currency?: string | null;
    start?: number | null;
    final?: number | null;
    avg?: number | null;
    avgList?: number[] | null;
  };
  customsDuty?: {
    individual?: number | null;
    legalEntity?: number | null;
  };
  disposalFee?: {
    individual?: number | null;
    legalEntity?: number | null;
  };
  saleCountry?: ('JAPAN' | 'CHINA') | null;
  engineType?: ('gasoline' | 'diesel' | 'electric' | 'hybrid-gasoline' | 'hybrid-diesel') | null;
  mileageKm?: number | null;
  date: string;
  color?: string | null;
  saleStatus?: string | null;
  isSold?: boolean | null;
  isSanctions?: boolean | null;
  isFinish?: boolean | null;
  updatedAt: string;
  createdAt: string;
}